+++ recipes/WSJ0Mix/separation/train/train.py
@@ -530,7 +530,7 @@ if __name__ == "__main__":
         sys.exit(1)
 
     # Data preparation
-    from recipes.WSJ2Mix.prepare_data import prepare_wsjmix  # noqa
+    from recipes.WSJ0Mix.prepare_data import prepare_wsjmix  # noqa
 
     run_on_main(
         prepare_wsjmix,
